requirements.txt
Flask==0.10.1
Jinja2==2.7.3
uWSGI==2.0.11.2
requests==2.7.0
Pillow==3.2.0
